requirements.txt
re
cowsay
csv
sys
pytest
Figlet
datetime
mock
builtins